+++ Tracker-ios/OnboardingScreen/ViewModel/OnboardingScreenViewModel.swift
@@ -1,25 +1,19 @@
 import UIKit
 
-// MARK: - OnboardingScreenPresenter
-final class OnboardingScreenPresenter {
+// MARK: - OnboardingScreenViewModel
+final class OnboardingScreenViewModel {
 
     // MARK: - Properties and Initializers
-    private weak var viewController: OnboardingScreenController?
     private let pages: [UIViewController] = [
         OnboardingPageController(text: "MONITOR_WHAT_YOU_NEED".localized,
                                  backgroundImage: UIImage(named: K.ImageNames.firstOnboardingImage) ?? UIImage()),
         OnboardingPageController(text: "EVEN_IF_THIS_IS_NOT_YOGA".localized,
                                  backgroundImage: UIImage(named: K.ImageNames.secondOnboardingImage) ?? UIImage())
     ]
-
-    init(viewController: OnboardingScreenController? = nil) {
-        self.viewController = viewController
-        viewController?.onboardingScreenView.delegate = self
-    }
 }
 
 // MARK: - Helpers
-extension OnboardingScreenPresenter {
+extension OnboardingScreenViewModel {
 
     func giveFirstPage() -> UIViewController? {
         pages.first
@@ -50,14 +44,8 @@ extension OnboardingScreenPresenter {
     func giveIndex(for currentViewController: UIViewController) -> Int? {
         pages.firstIndex(of: currentViewController)
     }
-}
 
-// MARK: - OnboardingScreenViewDelegate
-extension OnboardingScreenPresenter: OnboardingScreenViewDelegate {
-
-    func acceptAndProceedToMainScreen() {
+    func acceptOnboarding() {
         UserDefaultsManager.shared.saveOnboardingState()
-        UIApplication.shared.windows.first?.rootViewController = TabBarController()
-        UIApplication.shared.windows.first?.makeKeyAndVisible()
     }
 }
