+++ Tracker-ios/TrackerCategoryScreen/ViewModel/TrackerCategoryScreenViewModel.swift
@@ -1,10 +1,3 @@
-//
-//  TrackerCategoryScreenViewModel.swift
-//  Tracker-ios
-//
-//  Created by Илья Валито on 22.04.2023.
-//
-
 import UIKit
 
 // MARK: TrackerCategoryScreenViewModel
